+++ scripts/solo_agents_gi.py
@@ -40,7 +40,6 @@ from nip.utils.experiments import (
 )
 
 MULTIPROCESS = True
-TEST_SIZE = 0.2
 
 param_grid = dict(
     dataset_name=["eru10000"],